+++ captum/optim/_utils/image/dataset.py
