--- captum/optim/_utils/images.py
@@ -1,18 +1,0 @@
-from typing import Optional, Tuple
-
-
-def get_neuron_pos(
-    H: int, W: int, x: Optional[int] = None, y: Optional[int] = None
-) -> Tuple[int, int]:
-    if x is None:
-        _x = W // 2
-    else:
-        assert x < W
-        _x = x
-
-    if y is None:
-        _y = H // 2
-    else:
-        assert y < H
-        _y = y
-    return _x, _y
