+++ tests/optim/helpers/numpy_common.py
