+++ tests/optim/utils/image/__init__.py
@@ -1,0 +1,1 @@
+# Image utils tests
