--- tests/optim/utils/test_images_utils.py
@@ -1,38 +1,0 @@
-#!/usr/bin/env python3
-import unittest
-
-from captum.optim._utils.images import get_neuron_pos
-
-
-class TestGetNeuronPos(unittest.TestCase):
-    def test_get_neuron_pos_hw(self) -> None:
-        W, H = 128, 128
-        x, y = get_neuron_pos(H, W)
-
-        self.assertEqual(x, W // 2)
-        self.assertEqual(y, H // 2)
-
-    def test_get_neuron_pos_xy(self) -> None:
-        W, H = 128, 128
-        x, y = get_neuron_pos(H, W, 5, 5)
-
-        self.assertEqual(x, 5)
-        self.assertEqual(y, 5)
-
-    def test_get_neuron_pos_x_none(self) -> None:
-        W, H = 128, 128
-        x, y = get_neuron_pos(H, W, 5, None)
-
-        self.assertEqual(x, 5)
-        self.assertEqual(y, H // 2)
-
-    def test_get_neuron_pos_none_y(self) -> None:
-        W, H = 128, 128
-        x, y = get_neuron_pos(H, W, None, 5)
-
-        self.assertEqual(x, W // 2)
-        self.assertEqual(y, 5)
-
-
-if __name__ == "__main__":
-    unittest.main()
